Update TCPConfig struct with u64 timeout value

+++ cs144_rust/src/tcp_helpers/tcp_config.rs
@@ -2,13 +2,13 @@ use crate::{util::address::Address, wrapping_integers::WrappingInt32};
 
 pub const DEFAULT_CAPACITY: usize = 64000; // 默认容量
 pub const MAX_PAYLOAD_SIZE: usize = 1452; // 适合在IPv4或UDP数据报中的最大TCP负载
-pub const TIMEOUT_DFLT: u16 = 1000; // 默认的重传超时时间为 1 秒 1000ms
+pub const TIMEOUT_DFLT: u64 = 1000; // 默认的重传超时时间为 1 秒 1000ms
 pub const MAX_RETX_ATTEMPTS: usize = 8; // 放弃之前的最大重传尝试次数
 
 /// Config for TCP sender and receiver
 #[allow(dead_code)]
 pub struct TCPConfig {
-    pub rt_timeout: u16,                  // 重传超时的初始值，以毫秒为单位
+    pub rt_timeout: u64,                  // 重传超时的初始值，以毫秒为单位
     pub recv_capacity: usize,             // 接收容量，以字节为单位
     pub send_capacity: usize,             // 发送容量，以字节为单位
     pub fixed_isn: Option<WrappingInt32>, // 可选的固定初始序列号
